requirements.txt
async-lru==1.0.2
FormEncode==2.0.0
peewee==3.14.4
pyaes==1.6.1
pydantic==1.8.2
PyDispatcher==2.0.5
Pyrogram==1.2.9
PySocks==1.7.1
python-dotenv==0.19.0
six==1.16.0
SQLObject==3.9.1
typing-extensions==3.10.0.0
